# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MyGram is a free photo sharing app written in Go. People can share, view, and comment photos by everyone. Anyone can create an account by registering an email address and creating a username.",
        "title": "Fp-myGram",
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/socialmedias": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all social media with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Fetch all social media",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataFetchedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create and store a social media with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Add a social media",
                "parameters": [
                    {
                        "description": "Add Social Media",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.AddSocialMedia"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataAddedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/socialmedias/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a social media by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Update a social media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SocialMedia ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Social Media",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.UpdateSocialMedia"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataUpdatedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a social media by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "socialmedias"
                ],
                "summary": "Delete a social media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SocialMedia ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseMessageDeletedSocialMedia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_socialmedia_utils.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mygram-api_socialmedia_utils.ResponseMessage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "the error explained here"
                },
                "status": {
                    "type": "string",
                    "example": "fail"
                }
            }
        },
...
        "utils.AddSocialMedia": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                }
            }
        },
...
        "utils.AddedSocialMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "the created at generated here"
                },
                "id": {
                    "type": "string",
                    "example": "the social media id generated here"
                },
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
...
        "utils.ResponseDataAddedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.AddedSocialMedia"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseDataFetchedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.SocialMedias"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseDataUpdatedSocialMedia": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.UpdatedSocialMedia"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseMessageDeletedSocialMedia": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "your social media has been successfully deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.SocialMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "here is the generated created at"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated social media id"
                },
                "name": {
                    "type": "string",
                    "example": "Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.example.com/johndoe"
                },
                "updated_at": {
                    "type": "string",
                    "example": "here is the generated updated at"
                },
                "user": {
                    "$ref": "#/definitions/utils.User"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
        "utils.SocialMedias": {
            "type": "object",
            "properties": {
                "social_medias": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.SocialMedia"
                    }
                }
            }
        },
...
        "utils.UpdateSocialMedia": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "New Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.newexample.com/johndoe"
                }
            }
        },
...
        "utils.UpdatedSocialMedia": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "here is the generated social media id"
                },
                "name": {
                    "type": "string",
                    "example": "New Example"
                },
                "social_media_url": {
                    "type": "string",
                    "example": "https://www.newexample.com/johndoe"
                },
                "updated_at": {
                    "type": "string",
                    "example": "the updated at generated here"
                },
                "user_id": {
                    "type": "string",
                    "example": "here is the generated user id"
                }
            }
        },
...
        "utils.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated user id"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}